Move the ItemDataOps into the shared SierraQueryOps trait
