+++ pipeline/transformer/transformer_sierra/src/main/scala/uk/ac/wellcome/platform/transformer/sierra/transformers/SierraProduction.scala
@@ -1,7 +1,6 @@
 package uk.ac.wellcome.platform.transformer.sierra.transformers
 
 import uk.ac.wellcome.platform.transformer.sierra.exceptions.CataloguingException
-import uk.ac.wellcome.platform.transformer.sierra.source._
 import uk.ac.wellcome.platform.transformer.sierra.transformers.parsers.Marc008Parser
 import weco.catalogue.internal_model.identifiers.IdState
 import weco.catalogue.internal_model.work._
